Aplicando notificação

+++ src/Business/Services/FornecedorService.cs
@@ -9,10 +9,14 @@ public class FornecedorService(IFornecedorRepository fornecedorRepository) : Bas
 {
     public async Task Adicionar(Fornecedor fornecedor)
     {
-        
+
         if (!ExecuteValidacao(new FornecedorValidation(), fornecedor) || !ExecuteValidacao(new EnderecoValidation(), fornecedor.Endereco)) return;
 
-        // validar se ja existe um fornecedor com o mesmo documento;
+        if (fornecedorRepository.Buscar(f => f.Documento == fornecedor.Documento).Result.Any())
+        {
+            Notificar("Já existe um fornecedor com este documento informado.");
+            return;
+        }
 
         await fornecedorRepository.Adicionar(fornecedor);
     }
@@ -21,11 +25,38 @@ public class FornecedorService(IFornecedorRepository fornecedorRepository) : Bas
     {
         if (!ExecuteValidacao(new FornecedorValidation(), fornecedor)) return;
 
+        if (fornecedorRepository.Buscar(f => f.Documento == fornecedor.Documento && f.Id != fornecedor.Id).Result.Any())
+        {
+            Notificar("Já existe um fornecedor com este documento informado.");
+            return;
+        }
+
         await fornecedorRepository.Atualizar(fornecedor);
     }
 
     public async Task Remover(Guid id)
     {
+        var fornecedor = await fornecedorRepository.ObterFornecedorProdutosEndereco(id);
+
+        if (fornecedor == null)
+        {
+            Notificar("Fornecedor não encontrado.");
+            return;
+        }
+
+        if (fornecedor.Produtos.Any())
+        {
+            Notificar("O fornecedor possui produtos cadastrados!");
+            return;
+        }
+
+        var endereco = await fornecedorRepository.ObterEnderecoPorFornecedor(id);
+
+        if (endereco != null)
+        {
+            await fornecedorRepository.RemoverEnderecoFornecedor(endereco);
+        }
+
         await fornecedorRepository.Remover(id);
     }
